[enclave] move verify_worker_responses to onchain storage mod.
